requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
apache-airflow==2.7.3
apache-airflow-providers-common-sql==1.15.0
apache-airflow-providers-ftp==3.10.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-sqlite==3.8.2
apispec==6.8.2
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
cattrs==25.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.0
coverage==7.10.6
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==45.0.6
cycler==0.12.1
databricks-cli==0.18.0
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docker==6.1.3
durationpy==0.10
dynaconf==3.2.11
email-validator==1.3.1
entrypoints==0.4
evidently==0.2.2
exceptiongroup==1.3.0
Faker==37.6.0
fastapi==0.103.1
filelock==3.19.1
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.11.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-re2==1.1.20250805
googleapis-common-protos==1.70.0
graphviz==0.21
grpcio==1.74.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
kubernetes==33.1.0
lazy-object-proxy==1.12.0
limits==4.2
linkify-it-py==2.0.3
litestar==2.17.0
litestar-htmx==0.5.0
lockfile==0.12.2
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.26.1
matplotlib==3.9.4
mdit-py-plugins==0.4.2
mdurl==0.1.2
mlflow==2.7.1
more-itertools==10.8.0
msgspec==0.19.0
multidict==6.6.4
multipart==1.3.0
mypy_extensions==1.1.0
narwhals==2.2.0
nltk==3.8.1
numpy==1.23.5
oauthlib==3.3.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
ordered-set==4.1.0
packaging==23.2
pandas==1.5.3
pathspec==0.12.1
patsy==1.0.1
pendulum==2.1.2
Pillow==9.5.0
plotly==5.24.1
pluggy==1.6.0
polyfactory==2.22.2
prison==0.2.1
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
pyaml==25.7.0
pyarrow==13.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.4.2
pytest-cov==4.1.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2023.4
pytzdata==2020.1
PyYAML==6.0.2
querystring-parser==1.2.4
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.1
rich-click==1.8.9
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.3.0
scikit-optimize==0.9.0
scipy==1.13.1
seaborn==0.13.2
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
starlette==0.27.0
statsmodels==0.14.5
streamlit==1.27.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
typer==0.17.3
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
ujson==5.11.0
unicodecsv==0.14.1
urllib3==1.26.20
uuid6==2025.0.1
uvicorn==0.23.2
uvloop==0.21.0
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.2.3
wrapt==1.17.3
WTForms==3.0.1
xgboost==1.7.6
yarl==1.20.1
zipp==3.23.0
